requirements.txt
Django>=3.0,<4.0
argon2-cffi>=20.1.0
djangorestframework>=3.12.2